requirements.txt
altair==4.2.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
entrypoints==0.4
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.6.0
Jinja2==3.1.2
jsonschema==4.17.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
numpy~=1.24.3
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
protobuf==3.20.3
pyarrow==12.0.0
pydeck==0.8.0
Pygments==2.15.1
Pympler==1.0.1
PyMySQL==1.0.3
pyrsistent==0.19.3
pysqlite3-wheels==0.5.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
rich==13.3.5
six==1.16.0
smmap==5.0.0
streamlit~=1.22.0
streamlit-option-menu==0.3.6
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
typing_extensions==4.6.2
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.2
validators==0.20.0
watchdog==3.0.0
zipp==3.15.0
tensorflow==2.12.0
plotly==5.14.1
matplotlib~=3.7.1
folium~=0.14.0
streamlit-chat==0.0.2.2
streamlit-folium==0.12.0
seaborn==0.12.2
keras~=2.12.0